// lib/api/generated/models/analysisResultResponseDataSectionsItemKind.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Heymoa API
 * Heymoa 서버 REST API
 * OpenAPI spec version: 1.0.0
 */

/**
 * 항목 종류
 */
export type AnalysisResultResponseDataSectionsItemKind =
  (typeof AnalysisResultResponseDataSectionsItemKind)[keyof typeof AnalysisResultResponseDataSectionsItemKind];

export const AnalysisResultResponseDataSectionsItemKind = {
  OVERVIEW: "OVERVIEW",
  ACTION_ITEM: "ACTION_ITEM",
  DECISION: "DECISION",
} as const;